backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
redis==5.0.1
celery==5.3.4
websockets==12.0
openai==1.3.7
anthropic==0.7.7
sentence-transformers==2.2.2
numpy==1.24.3
python-dotenv==1.0.0
PyJWT==2.8.0
email-validator==2.1.0
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2
asyncpg==0.29.0